feat: update integer type definitions for 64-bit integers
Change I64 and U64 types from long to long long for better compatibility and standard compliance. This ensures consistent behavior across different platforms.

+++ orbit/datatype.h
@@ -18,8 +18,8 @@ typedef unsigned short U16;
 typedef int I32;
 typedef unsigned int U32;
 
-typedef long I64;
-typedef unsigned long U64;
+typedef long long I64;
+typedef unsigned long long U64;
 
 typedef U64 PtrSize;
 typedef U64 MSize;
